Function rust_template::greet

pub fn greet()

Rust Repository Template 🦀

Repository template to get quickly started with writing Rust libraries, ready for distributing.

Features

Remote development support with GitHub Codespaces

Debugging with the LLDB Debugger tool in Visual Studio Code (VSCode Marketplace)

Fuzz testing with LLVM’s libFuzzer tool and cargo fuzz (Reference)

Performance benchmarks in Rust with Criterion and Iai (suitable to run in GitHub Actions CI environments)

CI/CD support with GitHub Actions, allowing to automate:

Running tests and benchmarks

Running Rustfmt and Clippy for detecting formatting and linting errors, respectively

Daily, midnight scheduled security audits of Rust packages with actions-rs/audit-check

Generating performance benchmark graphs, auto-publishing to GitHub Pages

Documentation of API docs and mdBook using GitHub Pages

Semantic version bumping, CHANGELOG.md updates, and new package releases

Configure

GitHub Codespaces: devcontainer.json (Reference)

GitHub Action workflows: .github/workflows (Reference)

Cargo package: Cargo.toml (Reference)

Clippy (Rust linter): .clippy.toml (Repository, Reference)

Rustfmt (Rust formatter): .rustfmt.toml (Repository, Reference)

{{library}}

{{library description}}

Install

cargo add {{library}}

License

This software is openly licensed and distributed under the MIT License.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the MIT license, shall be licensed as above, without any additional terms or conditions. Say hi to the entire world :)